blender/pyproject.toml
[tool.poetry]
name = "blender"
version = "0.1.0"
description = "Calculate cross section area for SRP"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
pillow = "^9.5.0"
numpy = "^1.25.0"
matplotlib = "^3.7.1"
scipy = "^1.10.1"
pyquaternion = "^0.9.9"
skyfield = "^1.46"
astropy = "^5.3"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.1"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
